requirements.txt
pytesseract==0.3.10
Pillow==10.4.0
SQLAlchemy==1.4.39
robotframework-seleniumlibrary==6.5.0
selenium==4.23.1
selenium-wire==5.1.0
webdriver-manager==4.0.2
